Uptodate v0.3

+++ src/main/java/me/artemiyulyanov/uptodate/minio/MinioService.java
@@ -3,16 +3,23 @@ package me.artemiyulyanov.uptodate.minio;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.*;
 import jakarta.annotation.PostConstruct;
+import me.artemiyulyanov.uptodate.models.Article;
+import me.artemiyulyanov.uptodate.models.ArticleComment;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.util.List;
 
 @Service
 public class MinioService {
+    public static final String ARTICLE_RESOURCES_FOLDER = "/articles/%d";
+    public static final String ARTICLE_COMMENT_RESOURCES_FOLDER = "/articles/%d/comments/%d";
+
     @Autowired
     private AmazonS3 amazonS3;
 
@@ -26,19 +33,29 @@ public class MinioService {
         }
     }
 
-    public void uploadFile(String objectKey, MultipartFile file) throws IOException {
-        InputStream inputStream = file.getInputStream();
-        long contentLength = file.getSize();
+    public boolean uploadFile(String objectKey, MultipartFile file) {
+        if(amazonS3.doesObjectExist(bucket, objectKey)) return false;
 
-        PutObjectRequest putObjectRequest = new PutObjectRequest(
-                bucket,
-                objectKey,
-                inputStream,
-                new ObjectMetadata()
-        );
-        putObjectRequest.getMetadata().setContentLength(contentLength);
+        try (InputStream inputStream = file.getInputStream()) {
+            long contentLength = file.getSize();
 
-        amazonS3.putObject(putObjectRequest.withCannedAcl(CannedAccessControlList.PublicRead));
+            PutObjectRequest putObjectRequest = new PutObjectRequest(
+                    bucket,
+                    objectKey,
+                    inputStream,
+                    new ObjectMetadata()
+            );
+            putObjectRequest.getMetadata().setContentLength(contentLength);
+
+            amazonS3.putObject(putObjectRequest.withCannedAcl(CannedAccessControlList.PublicRead));
+            return true;
+        } catch (IOException e) {
+            return false;
+        }
+    }
+
+    public void removeFile(String objectKey) {
+        if(amazonS3.doesObjectExist(bucket, objectKey)) amazonS3.deleteObject(new DeleteObjectRequest(bucket, objectKey));
     }
 
     public MinioMediaFile getFile(String objectKey) {
@@ -49,5 +66,23 @@ public class MinioService {
                 .inputStream(s3Object.getObjectContent())
                 .objectKey(objectKey)
                 .build();
+    }
+
+    public void saveArticleResources(Article article, List<MultipartFile> resources) {
+        resources.forEach(file -> uploadFile(String.format(ARTICLE_RESOURCES_FOLDER, article.getId()) + File.separator + file.getOriginalFilename(), file));
+    }
+
+    public void saveArticleCommentResources(ArticleComment comment, List<MultipartFile> resources) {
+        resources.forEach(file -> uploadFile(String.format(ARTICLE_COMMENT_RESOURCES_FOLDER, comment.getArticle().getId(), comment.getId()) + File.separator + file.getOriginalFilename(), file));
+    }
+
+    public void deleteArticleResources(Article article) {
+        String resourcesFolder = String.format(ARTICLE_RESOURCES_FOLDER, article.getId());
+        if(amazonS3.doesObjectExist(bucket, resourcesFolder)) removeFile(resourcesFolder);
+    }
+
+    public void deleteArticleCommentResources(ArticleComment comment) {
+        String resourcesFolder = String.format(ARTICLE_RESOURCES_FOLDER, comment.getAuthor().getId(), comment.getId());
+        if(amazonS3.doesObjectExist(bucket, resourcesFolder)) removeFile(resourcesFolder);
     }
 }
